Updates for pipeline integrations
