+++ shelvery_tests/conftest.py
@@ -4,15 +4,31 @@ import os
 import time
 from shelvery.aws_helper import AwsHelper
 
-from shelvery_tests.test_functions import cleanupSnapshots
+from shelvery_tests.cleanup_functions import cleanupSnapshots
 
+source_account = None
+destination_account = None
+
+def pytest_addoption(parser):
+    parser.addoption("--destination", action="store", default="None")
+    parser.addoption("--source", action="store", default="None")
+
+
+def pytest_configure(config):
+    global source_account
+    global destination_account
+    
+    source_account = config.getoption('--source')
+    destination_account = config.getoption('--destination')
+
+        
 @pytest.fixture(scope="session", autouse=True)
 def setup(request):
 
     sts = AwsHelper.boto3_client('sts')
     id = str(sts.get_caller_identity()['Account'])
 
-    if id == '988966687271':
+    if id == source_account:
 
         cfclient = boto3.client('cloudformation')
         cwd = os.getcwd()
