+++ shelvery_tests/ec2ami_integration_test.py
@@ -26,6 +26,7 @@ from shelvery.engine import S3_DATA_PREFIX
 from shelvery.runtime_config import RuntimeConfig
 from shelvery.backup_resource import BackupResource
 from shelvery.aws_helper import AwsHelper
+from shelvery_tests.conftest import destination_account
 
 print(f"Python lib path:\n{sys.path}")
 
@@ -66,7 +67,7 @@ class ShelveryEC2AmiIntegrationTestCase(unittest.TestCase):
 
         createBackupTags(ec2client,[instance_id],"shelvery-test-ec2")
 
-        self.share_with_id = 186991632813
+        self.share_with_id = destination_account
 
     @pytest.mark.source
     def test_Cleanup(self):
