+++ shelvery_tests/rds_integration_test.py
@@ -27,6 +27,7 @@ from shelvery.engine import S3_DATA_PREFIX
 from shelvery.runtime_config import RuntimeConfig
 from shelvery.backup_resource import BackupResource
 from shelvery.aws_helper import AwsHelper
+from shelvery_tests.conftest import destination_account
 
 print(f"Python lib path:\n{sys.path}")
 
@@ -57,7 +58,7 @@ class ShelveryRDSIntegrationTestCase(unittest.TestCase):
                       rdsinstance['DBInstanceArn'],
                       "shelvery-test-rds")
 
-        self.share_with_id = 186991632813
+        self.share_with_id = destination_account
 
     @pytest.mark.source
     def test_Cleanup(self):
